+++ headers_test.go
@@ -13,21 +13,21 @@ func TestHeadersContext(t *testing.T) {
 
 	th := make(TrustedHeaders)
 
-	th[SFHeaderUserID] = "my-user-id"
+	th[HeaderUserID] = "my-user-id"
 	th["random-key"] = "102"
 
 	ctx = WithTrustedHeaders(ctx, th)
 
 	got := FromContext(ctx)
 
-	assert.Equal(t, "my-user-id", got.Get(SFHeaderUserID))
+	assert.Equal(t, "my-user-id", got.Get(HeaderUserID))
 	assert.Equal(t, "102", got.Get("random-key"))
 }
 
 func TestHeadersMetadataContext(t *testing.T) {
 	th := make(TrustedHeaders)
-	th[SFHeaderUserID] = "my-user-id"
-	th[SFHeaderIP] = "10.2.3.4"
+	th[HeaderUserID] = "my-user-id"
+	th[HeaderIP] = "10.2.3.4"
 
 	ctx := context.Background()
 	ctx = th.ToOutgoingGRPCContext(ctx)
@@ -35,7 +35,7 @@ func TestHeadersMetadataContext(t *testing.T) {
 	md, ok := metadata.FromOutgoingContext(ctx)
 	assert.True(t, ok)
 
-	assert.Equal(t, []string{"my-user-id"}, md.Get(SFHeaderUserID))
+	assert.Equal(t, []string{"my-user-id"}, md.Get(HeaderUserID))
 }
 
 func TestNilDoesntPanic(t *testing.T) {
@@ -46,5 +46,6 @@ func TestNilDoesntPanic(t *testing.T) {
 	assert.Equal(t, "", h.RealIP())
 	assert.Equal(t, "", h.UserID())
 	assert.Equal(t, "", h.Meta())
+	assert.Equal(t, "", h.PlanTier())
 	assert.Equal(t, "", h.Get("something"))
 }