+++ middleware/http/middleware_test.go
@@ -22,8 +22,8 @@ func (t testAuthenticators) Authenticate(ctx context.Context, path string, heade
 	for key, values := range headers {
 		out[key] = values[0]
 	}
-	out["x-sf-substreams-ll"] = "987"
-	out["x-sf-user-id"] = "a1b2c3"
+	out["x-substreams-ll"] = "987"
+	out["x-user-id"] = "a1b2c3"
 
 	return dauth.WithTrustedHeaders(ctx, out), nil
 }
@@ -33,7 +33,7 @@ func TestAuthMiddleware_validateAuth(t *testing.T) {
 	require.NoError(t, err)
 
 	request.Header.Set("authorization", "bearer jwt_token")
-	request.Header.Set("X-SF-SUBSTREAMS-LL", "123")
+	request.Header.Set("X-SUBSTREAMS-LL", "123")
 
 	authenticator := &testAuthenticators{}
 	newRequest, err := validateAuth(request, authenticator)
@@ -41,6 +41,6 @@ func TestAuthMiddleware_validateAuth(t *testing.T) {
 
 	trusted := dauth.FromContext(newRequest.Context())
 
-	assert.Equal(t, "987", trusted.Get("x-sf-substreams-ll"))
-	assert.Equal(t, "a1b2c3", trusted.Get("X-Sf-User-ID"))
+	assert.Equal(t, "987", trusted.Get("x-substreams-ll"))
+	assert.Equal(t, "a1b2c3", trusted.Get("X-User-ID"))
 }