+++ trust/trust_test.go
@@ -2,8 +2,9 @@ package trust
 
 import (
 	"context"
-	"go.uber.org/zap"
 	"testing"
+
+	"go.uber.org/zap"
 
 	"github.com/streamingfast/dauth"
 	"github.com/stretchr/testify/assert"
@@ -12,13 +13,13 @@ import (
 
 func TestParse(t *testing.T) {
 	Register()
-	a, err := dauth.New("trust://?allowed=x-sf-something,x-sf-SoMeThingElse", zap.NewNop())
+	a, err := dauth.New("trust://?allowed=x-something,x-SoMeThingElse", zap.NewNop())
 	require.NoError(t, err)
 
 	p := a.(*trustPlugin)
 	assert.Equal(t, map[string]bool{
-		"x-sf-something":     true,
-		"x-sf-somethingelse": true,
+		"x-something":     true,
+		"x-somethingelse": true,
 	},
 		p.allowed)
 
@@ -29,33 +30,33 @@ func TestAuthenticateEmpty(t *testing.T) {
 	p := &trustPlugin{}
 
 	ctx, err := p.Authenticate(context.Background(), "", map[string][]string{
-		"x-sf-something":     []string{"someval", "ignored"},
-		"x-sf-somethingelse": []string{"someotherval", "ignored"},
+		"x-something":     []string{"someval", "ignored"},
+		"x-somethingelse": []string{"someotherval", "ignored"},
 	}, "10.0.0.1")
 	require.NoError(t, err)
 
 	auth := dauth.FromContext(ctx)
-	assert.Equal(t, "someval", auth.Get("x-sf-something"))
-	assert.Equal(t, "someotherval", auth.Get("x-sf-somethingelse"))
+	assert.Equal(t, "someval", auth.Get("x-something"))
+	assert.Equal(t, "someotherval", auth.Get("x-somethingelse"))
 }
 
 func TestAuthenticateAllowed(t *testing.T) {
 
 	p := &trustPlugin{
 		allowed: map[string]bool{
-			"x-sf-something":     true,
-			"x-sf-somethingelse": true,
+			"x-something":     true,
+			"x-somethingelse": true,
 		},
 	}
 
 	ctx, err := p.Authenticate(context.Background(), "", map[string][]string{
-		"x-sf-something": []string{"someval", "ignored"},
-		"x-sf-forbidden": []string{"forbiddenval", "ignored"},
+		"x-something": []string{"someval", "ignored"},
+		"x-forbidden": []string{"forbiddenval", "ignored"},
 	}, "10.0.0.1")
 	require.NoError(t, err)
 
 	auth := dauth.FromContext(ctx)
-	assert.Equal(t, "someval", auth.Get("x-sf-something"))
-	assert.Equal(t, "", auth.Get("x-sf-somethingelse"))
-	assert.Equal(t, "", auth.Get("x-sf-forbidden"))
+	assert.Equal(t, "someval", auth.Get("x-something"))
+	assert.Equal(t, "", auth.Get("x-somethingelse"))
+	assert.Equal(t, "", auth.Get("x-forbidden"))
 }